Move `asImmutable()` to internal interface
This reduces the API surface to elements which might be of interest for users.
